Move generic code to better location
