+++ src/internal/connector/graph/container_resolver_test.go
@@ -1,4 +1,4 @@
-package exchange
+package graph
 
 import (
 	"context"
